pyproject.toml
[project]
name = "flexcity-techtest"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-ninja (>=1.5.3,<2.0.0)",
    "ortools (>=9.15.6755,<10.0.0)",
    "black (>=26.3.0,<27.0.0)",
    "faker (>=40.8.0,<41.0.0)",
]

[tool.poetry]
packages = [{include = "flexcity_techtest", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
